numeral: change int to uint16

+++ roman-numerals/numeral.go
@@ -2,14 +2,14 @@ package main
 
 import "strings"
 
-func ConvertToArabic(roman string) (total int) {
+func ConvertToArabic(roman string) (total uint16) {
 	for _, symbols := range windowedRoman(roman).Symbols() {
 		total += allRomanNumerals.ValueOf(symbols...)
 	}
 	return
 }
 
-func ConvertToRoman(arabic int) string {
+func ConvertToRoman(arabic uint16) string {
 	var result strings.Builder
 
 	for _, numeral := range allRomanNumerals {
@@ -23,13 +23,13 @@ func ConvertToRoman(arabic int) string {
 }
 
 type RomanNumeral struct {
-	Value  int
+	Value  uint16
 	Symbol string
 }
 
 type romanNumerals []RomanNumeral
 
-func (r romanNumerals) ValueOf(symbols ...byte) int {
+func (r romanNumerals) ValueOf(symbols ...byte) uint16 {
 	symbol := string(symbols)
 	for _, s := range r {
 		if s.Symbol == symbol {
@@ -83,6 +83,6 @@ func (w windowedRoman) Symbols() (symbols [][]byte) {
 	return
 }
 
-func isSubtractive(symbol uint8) bool {
+func isSubtractive(symbol byte) bool {
 	return symbol == 'I' || symbol == 'X' || symbol == 'C'
 }
